+++ fiddles/bash/fiddle-0100-ReactCreateApp/CHANGELOG.md
@@ -8,3 +8,4 @@
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
+* Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)